+++ History.md
@@ -1,5 +1,6 @@
 2.0.10 / 2020-??-??
 ==================
+  * Replace 'eval' with 'JSON.parse' (https://github.com/tntim96/JSCover/issues/290)
   * Upgrade closure-compiler v20200101 to v20200628, Rhino 1.7.11 to 1.7.12
   * Internal: HtmlUnit 2.36 to 2.41.0, Mockito 3.2.4 to 3.3.3
   * Replace deprecated Maven Ant Tasks with Maven Artifact Resolver Ant Tasks (https://github.com/tntim96/JSCover/issues/285)